- Added an implementation of the OrphanPool - Added tests for the OrphanPool - Added contains function to MerkleMountainRange - Cleaned TxPriority and added get_fee_per_byte function to transaction - Modified the create_test_tx, create_test_block, create_test_kernel functions so that they can be used for the OrphanPool Tests - Added create_test_utxos and extend_test_utxos test_utils functions

+++ base_layer/core/src/mempool/unconfirmed_pool/unconfirmed_pool_storage.rs
@@ -102,9 +102,9 @@ impl UnconfirmedPoolStorage {
     }
 
     /// Returns a set of the highest priority unconfirmed transactions, that can be included in a block.
-    pub fn highest_priority_txs(&self, total_weight: usize) -> Result<Vec<Arc<Transaction>>, UnconfirmedPoolError> {
+    pub fn highest_priority_txs(&self, total_weight: u64) -> Result<Vec<Arc<Transaction>>, UnconfirmedPoolError> {
         let mut selected_txs: Vec<Arc<Transaction>> = Vec::new();
-        let mut curr_weight: usize = 0;
+        let mut curr_weight: u64 = 0;
         let mut curr_skip_count: usize = 0;
         for (_, tx_key) in self.txs_by_priority.iter().rev() {
             let ptx = self
